Relax patching checks for non-bootable disks (#1)
Changed the patching to support data disks, which still require a boot
sector DPB to work with a previously booted patched version of CP/M 3.

We now check the boot sector looks normal enough to patch, with either a
byte filler or a a common decrementing byte pattern. Only then do we add the
DPB to the end of the sector. This can still be overridden using the --force
option if the user decides it's safe.

We previously required both CPM3.SYS and CPMLDR to be present and patchable
before patching the boot sector. This has been relaxed to patch them only if
they are present.

The patched disk image is now only written if anything was patched.