Boot sector DPB is now a fallback option (#2)
Changed the patching to keep the original pre-track DPB reading, only trying
the boot sector if it wasn't found. If that fails we fall back on the
original CPM3.SYS behaviour of auto-detecting a DPB from the track format.
CPMLDR has no fallback and just hangs the system if there's no DPB.

The additional code requires use of some existing routines, which are found
using code matching that works with known versions of CPMLDR and CPM3.SYS.
The new code location is found using similar matching, after analysis of the
original code to determine a safe location.