Add support for overriding the default NABU DPB
CP/M 3 has a default internal DPB that is used for NABU disks that don't
have an explicit DPB. This only applies to disks used after CP/M 3 is
booted. CPMLDR still requires a track DPB, or a boot sector DPB on patched
disks.

A new -d option allows the internal default DPB to be overridden, so that it
defaults to a higher capacity format. If not specified the original internal
s4 format is used.